+++ src/pybritive/helpers/config.py
@@ -71,6 +71,8 @@ class ConfigManager:
         self.alias = None
         self.default_tenant = None
         self.tenants = None
+        self.tenants_by_name = None
+        self.aliases_and_names = None
         self.profile_aliases = {}
         self.cli = cli
         self.loaded = False
@@ -106,10 +108,17 @@ class ConfigManager:
         self.alias = None  # will be set in self.get_tenant()
         self.default_tenant = self.config.get('global', {}).get('default_tenant')
         self.tenants = {}
+        self.tenants_by_name = {}
         for key in self.config:
             if key.startswith('tenant-'):
                 alias = extract_tenant(key)
-                self.tenants[alias] = self.config[key]
+                item = {**self.config[key], **{'alias': alias}}
+                self.tenants[alias] = item
+
+                name = self.config[key].get('name', alias)
+                if name != alias:
+                    self.tenants_by_name[name] = item
+        self.aliases_and_names = {**self.tenants, **self.tenants_by_name}
         self.profile_aliases = self.config.get('profile-aliases', {})
         self.loaded = True
 
@@ -137,12 +146,12 @@ class ConfigManager:
             provided_tenant_name = list(self.tenants)[0]
 
         # if we get here then we now have a tenant name we can check to ensure exists
-        if provided_tenant_name not in self.tenants and not name:
+        if provided_tenant_name not in self.aliases_and_names and not name:
             raise click.ClickException(f'Tenant name "{provided_tenant_name}" not found in {self.path}')
 
         self.alias = provided_tenant_name or name
         # return details about the requested tenant
-        return self.tenants.get(provided_tenant_name, {'name': name})
+        return self.aliases_and_names.get(provided_tenant_name, {'name': name, 'alias': name})
 
     def save(self):
         self.validate()  # ensure we are actually writing a valid config